ApiFuncional

+++ APIProyecto/Models/IntencionDescripcion.cs
@@ -8,7 +8,7 @@ namespace APIProyecto.Models
         public int IdIntencionDescripcion { get; set; }
         
         public int Cantidad { get; set; }
-        public float PrecioTotal { get; set; }
+        public double PrecioTotal { get; set; }
         public bool Status {get;set;}
         // Clave foránea para la relación uno a muchos con Cliente
         public int ProductoColorTallaIdProductoColorTalla { get; set; }
